requirements.txt
python-pptx
Pillow
anthropic
streamlit
lxml
gspread
google-auth
